sm/app/flask_app/requirements.txt
flask>=2.2.5
flask-Limiter>=3.8.0
Werkzeug>=3.0.6
mysql-connector-python>=9.1.0
bcrypt==4.2.0
flask-bcrypt==1.0.1
PyJWT==2.9.0